Revert "refactor: :recycle: typescript for client"
This reverts commit fe1539bfef7a2081f01d5d94a639bbdf89313f3b.

+++ apps/client/src/stores/project-environment.ts
@@ -1,21 +1,26 @@
 import { defineStore } from 'pinia'
 import api from '@/api/index.js'
 import { useProjectStore } from '@/stores/project.js'
+import { EnvironmentModel, InitializeEnvironmentDto, UpdateEnvironmentDto } from '@dso-console/shared'
+import { projectMissing } from '@/utils/const'
 
 export const useProjectEnvironmentStore = defineStore('project-environment', () => {
   const projectStore = useProjectStore()
 
-  const addEnvironmentToProject = async (newEnvironment) => {
+  const addEnvironmentToProject = async (newEnvironment: InitializeEnvironmentDto['body']) => {
+    if (!projectStore.selectedProject) throw new Error(projectMissing)
     await api.addEnvironment(projectStore.selectedProject.id, newEnvironment)
     await projectStore.getUserProjects()
   }
 
-  const updateEnvironment = async (environment) => {
+  const updateEnvironment = async (environment: UpdateEnvironmentDto['body'] & { id: EnvironmentModel['id']}) => {
+    if (!projectStore.selectedProject) throw new Error(projectMissing)
     await api.updateEnvironment(projectStore.selectedProject.id, environment.id, environment)
     await projectStore.getUserProjects()
   }
 
-  const deleteEnvironment = async (environmentId) => {
+  const deleteEnvironment = async (environmentId: EnvironmentModel['id']) => {
+    if (!projectStore.selectedProject) throw new Error(projectMissing)
     await api.deleteEnvironment(projectStore.selectedProject.id, environmentId)
     await projectStore.getUserProjects()
   }
